+++ TD2/exo9/Vecteur3D.hpp
@@ -7,8 +7,8 @@ public :
 
 void affiche();
 void affiche(const char* message);
-Vecteur3D() : x(0), y(0), z(0){} // Constructeur en ligne
-Vecteur3D(int abs,int ord,int cot) : x(abs),y(ord),z(cot){} // Constructeur par défaut en ligne
+Vecteur3D() : x(0), y(0), z(0){} // Constructeur par défaut en ligne
+Vecteur3D(double abs,double ord,double cot) : x(abs),y(ord),z(cot){} // Constructeur en ligne
 
 private :
 double x,y,z;
